Move shared options-flattening helper to src/Shared
Relocate AtsOptionsFlattening from Aspire.TypeSystem to src/Shared/CodeGeneration
and consume it via <Compile Include> in both generators, instead of exposing it
through InternalsVisibleTo on the version-frozen contract assembly. Replace the
CoexistingCancellationTokenPolicy enum with a cancellationTokenIsSeparateParameter
bool so both values drive the single behavioral fork.
